Solves #10.
Small LLMs (like Mistral-7B) play fast and loose with quotes in strings so you can get single quoted strings (that is not JSON standard) and even a mix of different terminators in the same output. This is now supported and fixed.

Please note that something like 'string" will be fixed as 'string"' because otherwise the string "won't" will be broken.